+++ prost-types/Cargo.toml
@@ -18,13 +18,14 @@ doctest = false
 [features]
 default = ["std"]
 std = ["prost/std"]
+json = ["prost/json", "chrono"]
 
 [dependencies]
 base64 = "0.13"
 bytes = { version = "1", default-features = false }
 serde = { version = "1", features = ["derive"] }
-humantime = { version = "2.1" }
 prost = { version = "0.9.0", path = "..", default-features = false, features = ["prost-derive"] }
+chrono = { version = "0.4", optional = true }
 
 [dev-dependencies]
 proptest = "1"
